requirements.txt
openai-whisper
pyaudio
requests
django
djangorestframework
drf-yasg
pybars3
